[2단계 - 경로 조회 기능] 주노(최준호) 미션 제출합니다. (#105)
* docs: 요구사항 작성

* feat: db 환경분리

* refactor: 존재하지 않는 자원에 대해 수정, 삭제 수행 시 예외

* test: 구간을 구하는 시나리오테스트 작성

* feat: 경로탐색기능 추가

- 최단경로 탐색
- 요금 계산
- 경로 합 반환

* test: 최단경로 테스트 보완

* refactor: 경로 탐색에 사용되는 변수명 변경

- from, dest -> source, target

* refactor: 요금 용어 변경

- fee, money -> fare

* refactor: 요금 계산로직 수정 및 테스트 추가

* refactor: RouteFinder 추상화 및 Bean으로 등록

* test: 경로 탐색 예외 추가

* test: 최단거리 경로 탐색 추가

* refactor: Post 요청 시 URL만 반환하도록 수정

* refactor: Transactional Service 계층에서 전파하도록 수정

* test: post 요청 body 제거에 따른 인수테스트 수정

+++ src/main/java/subway/dao/H2LineDao.java
@@ -5,6 +5,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import subway.entity.LineEntity;
+import subway.exception.NoSuchLineException;
 
 import javax.sql.DataSource;
 import java.util.HashMap;
@@ -58,12 +59,23 @@ public class H2LineDao implements LineDao {
 
     @Override
     public void update(final LineEntity newLineEntity) {
+        if (!existsById(newLineEntity.getId())) {
+            throw new NoSuchLineException(newLineEntity.getId());
+        }
         String sql = "UPDATE line SET name = ?, color = ? WHERE id = ?";
         jdbcTemplate.update(sql, newLineEntity.getName(), newLineEntity.getColor(), newLineEntity.getId());
     }
 
     @Override
     public void deleteById(final Long id) {
+        if (!existsById(id)) {
+            throw new NoSuchLineException(id);
+        }
         jdbcTemplate.update("DELETE FROM line WHERE id = ?", id);
+    }
+
+    private boolean existsById(final Long id) {
+        String sql = "SELECT EXISTS(SELECT 1 FROM line WHERE id = ?)";
+        return jdbcTemplate.queryForObject(sql, new Object[]{id}, Boolean.class);
     }
 }
